Refacted prefilters since we're adding even more, added evidence_type, added study prefilter for rob visuals
